requirements.txt
python-dotenv>=1.0.0
google-generativeai>=0.3.1
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
pyttsx3>=2.90